requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
h5py==3.10.0
joblib==1.3.2
keras==2.12.0
numpy==1.23.5
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.31.0
scikit-learn==1.3.2